データのやりとり等に不都合なのでメイン画面のjsxを一つのコンポーネントにまとめた
