+++ kakeibo/resources/js/components/main.js
@@ -1,0 +1,153 @@
+import React from 'react';
+import axios from 'axios';
+
+class Main extends React.Component {
+    constructor(props) {
+        super(props);
+        let csrf_token = document.head.querySelector('meta[name="csrf-token"]').content;
+        let api_token = document.head.querySelector('meta[name="api-token"]').content;
+        this.state = {
+            title: '支出',
+            uri: '/main?param=e',
+            month_expenses: [],
+            month_incomes: [],
+            csrf_token: csrf_token,
+            api_token: api_token
+        };
+        this.setExpenseValue = this.setExpenseValue.bind(this);
+        this.setIncomeValue = this.setIncomeValue.bind(this);
+    }
+
+    componentDidMount() {
+        axios.get('api/main', {
+            params: {
+                api_token: this.state.api_token
+            }
+        })
+            .then((response) => {
+                this.setState({
+                    month_expenses: response.data.month_expenses,
+                    month_incomes: response.data.month_incomes
+                });
+            })
+            .catch((error) => {
+                console.log('error');
+            })
+            .then(() => {
+                console.log('always executed');
+            })
+    }
+
+    setExpenseValue() {
+        this.setState({
+            title: '支出',
+            uri: '/main?param=e'
+        });
+    }
+
+    setIncomeValue() {
+        this.setState({
+            title: '収入',
+            uri: '/main?param=i'
+        });
+    }
+
+    render() {
+        const { month_expenses, month_incomes } = this.state;
+        const month_expenses_block = month_expenses.map(month_expense => {
+            return (
+                <tbody>
+                    <tr>
+                        <td data-label="Name">{ month_expense.item }</td>
+                        <td data-label="Age">{ month_expense.amount }</td>
+                        <td data-label="Job">
+                            <form action={ '/main?param=e&id=' + month_expense.id } method='delete'>
+                                <button type="submit">削除</button>
+                            </form>
+                        </td>
+                    </tr>
+                </tbody>
+            );
+        });
+        const month_incomes_block = month_incomes.map(month_income => {
+            return (
+                <tbody>
+                    <tr>
+                        <td data-label="Name">{ month_income.item }</td>
+                        <td data-label="Age">{ month_income.amount }</td>
+                        <td data-label="Job">
+                            <form action={ '/main?param=i&id=' + month_income.id } method='delete'>
+                                <button type="submit">削除</button>
+                            </form>
+                        </td>
+                    </tr>
+                </tbody>
+            );
+        });
+
+        return (
+            <>
+                <div class="ui placeholder segment">
+                    <div class="ui icon header">
+                        {/* <h2>{{ $year }}年{{ $month }}月</h2> */}
+                        <div class="ui centered cards">
+                            <div class="card">
+                                <div class="content">
+                                    {/* <h1>{{ $month_total_amount }}</h1> */}
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <br />
+                <div class="ui placeholder segment">
+                    <div class="ui center aligned icon header">
+                        {/* TODO:入力欄とボタンを横並びにしたい(semantic uiのgrid?) */}
+                        <div className="ui buttons">
+                            <button className="ui button active" onClick={this.setExpenseValue}>支出</button>
+                            <div className="or"></div>
+                            <button className="ui positive button" onClick={this.setIncomeValue}>収入</button>
+                        </div>
+                        <form action={ this.state.uri } method="post">
+                            <input type="hidden" name="_token" value={ this.state.csrf_token } />
+                            <div className="ui huge form">
+                                <h3>{ this.state.title }</h3>
+                                <div className="two fields">
+                                    <div className="field">
+                                        <input type="text" name="item" placeholder="項目名" />
+                                    </div>
+                                    <div className="field">
+                                        <input type="number" name="amount" placeholder="金額" />
+                                    </div>
+                                </div>
+                                <button type="submit" className="positive large ui button">登録</button>
+                            </div>
+                        </form>
+                    </div>
+                </div>
+                <br />
+                <table className="ui celled table">
+                    <thead>
+                        <tr>
+                            <th>支出</th>
+                            <th>金額</th>
+                            <th></th>
+                        </tr>
+                    </thead>
+                    { month_expenses_block }
+                    <thead>
+                        <tr>
+                            <th>収入</th>
+                            <th>金額</th>
+                            <th></th>
+                        </tr>
+                    </thead>
+                    { month_incomes_block }
+                </table>
+                <br />
+            </>
+        );
+    }
+}
+
+export default Main;
